Refactored autonomous files, put it into its own folder
